removed warning
git-svn-id: https://client-sdk.forge.funambol.org/svn/client-sdk/branches/cpp-sdk/US1744@593 3966be30-755a-0410-815e-dd79f492ac3c

+++ src/cpp/common/client/FileSyncSource.cpp
@@ -381,7 +381,7 @@ bool FileSyncSource::filterOutgoingItem(const StringBuffer& fullName, struct sta
 
 bool FileSyncSource::checkFileExtension(const StringBuffer& fileName, const StringBuffer& extension)
 {
-    int pos = fileName.rfind(".");
+    unsigned long pos = fileName.rfind(".");
     
     if (pos == StringBuffer::npos) {
         return false;
